requirements.txt
langchain-core==0.2.43
langchain==0.2.16
langchain-community==0.2.16
langchain-pinecone==0.1.3
langchain-groq==0.1.10
sentence-transformers>=3.0.0
flask==3.0.3
pypdf==4.2.0
python-dotenv==1.0.1
ragas==0.1.21
datasets==2.19.2
numpy==1.26.4
pandas==2.1.4
gunicorn
rank_bm25>=0.2.2
pymupdf>=1.24.0
nltk>=3.8.1
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
requests>=2.31.0
-e .
